Admin check

+++ grammer.py
@@ -9,7 +9,8 @@ VERSION = "1.0"
 TOKEN_PATH = f"{system.scriptPath}/token.txt"
 START = f"{system.scriptPath}/data/message.txt"
 AUTOSTART = f"{system.scriptPath}/data/on_start.txt"
-ADMIN = f"{system.scriptPath}/admin.txt"
+ADMIN = int(file_io.readFile(f"{system.scriptPath}/admin.txt"))
+
 
 os.chmod(system.scriptPath, 777)
 
@@ -20,7 +21,7 @@ except:
 
 try:
     os.chdir(system.defaultStartFolder)
-    bot.send_message(int(file_io.readFile(ADMIN)), f"""PC Started
+    bot.send_message(ADMIN, f"""PC Started
     ==========
     Script path: `{system.scriptPath}`
     ==========
@@ -36,16 +37,22 @@ except:
 @bot.message_handler(commands=["start", "info", "help"])
 
 def start_message(message):
-    bot.send_message(message.chat.id, file_io.readFile(START), parse_mode="Markdown")
+    if message.chat.id == ADMIN:
+        bot.send_message(message.chat.id, file_io.readFile(START), parse_mode="Markdown")
+    else:
+        bot.send_message(message.chat.id, "Who are you?")
 
 @bot.message_handler(content_types=["text"])
 
 def text(message):
-    if message.text == "screen":
-        screen = system.getScreen()
-        bot.send_document(message.chat.id, open(screen, "rb"))
-        os.remove(screen)
+    if message.chat.id == ADMIN:
+        if message.text == "screen":
+            screen = system.getScreen()
+            bot.send_document(message.chat.id, open(screen, "rb"))
+            os.remove(screen)
+        else:
+            bot.send_message(message.chat.id, mes_io.messageProcessing(message.text), parse_mode="Markdown")
     else:
-        bot.send_message(message.chat.id, mes_io.messageProcessing(message.text), parse_mode="Markdown")
-
+        bot.send_message(message.chat.id, "Who are you?")
+        
 bot.polling(non_stop=True)
